+++ st2common/tests/unit/services/test_workflow.py
@@ -19,19 +19,19 @@ import mock
 
 from orquesta import exceptions as orquesta_exc
 from orquesta.specs import loader as specs_loader
-from orquesta import states as wf_lib_states
+from orquesta import states as wf_states
 
 import st2tests
 
 from st2common.bootstrap import actionsregistrar
 from st2common.bootstrap import runnersregistrar
-from st2common.exceptions import action as ac_exc
+from st2common.exceptions import action as action_exc
 from st2common.models.db import liveaction as lv_db_models
 from st2common.models.db import execution as ex_db_models
 from st2common.persistence import execution as ex_db_access
 from st2common.persistence import workflow as wf_db_access
-from st2common.services import action as ac_svc
-from st2common.services import workflows as wf_svc
+from st2common.services import action as action_service
+from st2common.services import workflows as workflow_service
 from st2common.transport import liveaction as lv_ac_xport
 from st2common.transport import publishers
 from st2tests.mocks import liveaction as mock_lv_ac_xport
@@ -81,12 +81,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Request the workflow execution.
         wf_def = self.get_wf_def(TEST_PACK_PATH, wf_meta)
         st2_ctx = self.mock_st2_context(ac_ex_db)
-        wf_ex_db = wf_svc.request(wf_def, ac_ex_db, st2_ctx)
+        wf_ex_db = workflow_service.request(wf_def, ac_ex_db, st2_ctx)
 
         # Check workflow execution is saved to the database.
         wf_ex_dbs = wf_db_access.WorkflowExecution.query(action_execution=str(ac_ex_db.id))
@@ -97,19 +97,19 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
         self.assertIsNotNone(wf_ex_db.id)
         self.assertGreater(wf_ex_db.rev, 0)
         self.assertEqual(wf_ex_db.action_execution, str(ac_ex_db.id))
-        self.assertEqual(wf_ex_db.status, wf_lib_states.REQUESTED)
+        self.assertEqual(wf_ex_db.status, wf_states.REQUESTED)
 
     def test_request_with_input(self):
         wf_meta = self.get_wf_fixture_meta_data(TEST_PACK_PATH, 'sequential.yaml')
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'], parameters={'who': 'stan'})
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Request the workflow execution.
         wf_def = self.get_wf_def(TEST_PACK_PATH, wf_meta)
         st2_ctx = self.mock_st2_context(ac_ex_db)
-        wf_ex_db = wf_svc.request(wf_def, ac_ex_db, st2_ctx)
+        wf_ex_db = workflow_service.request(wf_def, ac_ex_db, st2_ctx)
 
         # Check workflow execution is saved to the database.
         wf_ex_dbs = wf_db_access.WorkflowExecution.query(action_execution=str(ac_ex_db.id))
@@ -120,7 +120,7 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
         self.assertIsNotNone(wf_ex_db.id)
         self.assertGreater(wf_ex_db.rev, 0)
         self.assertEqual(wf_ex_db.action_execution, str(ac_ex_db.id))
-        self.assertEqual(wf_ex_db.status, wf_lib_states.REQUESTED)
+        self.assertEqual(wf_ex_db.status, wf_states.REQUESTED)
 
         # Check input and context.
         expected_input = {
@@ -137,8 +137,8 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Request the workflow execution.
         self.assertRaises(
-            ac_exc.InvalidActionReferencedException,
-            wf_svc.request,
+            action_exc.InvalidActionReferencedException,
+            workflow_service.request,
             self.get_wf_def(TEST_PACK_PATH, wf_meta),
             ac_ex_db,
             self.mock_st2_context(ac_ex_db)
@@ -149,12 +149,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Exception is expected on request of workflow execution.
         self.assertRaises(
             orquesta_exc.WorkflowInspectionError,
-            wf_svc.request,
+            workflow_service.request,
             self.get_wf_def(TEST_PACK_PATH, wf_meta),
             ac_ex_db,
             self.mock_st2_context(ac_ex_db)
@@ -165,12 +165,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Exception is expected on request of workflow execution.
         self.assertRaises(
             orquesta_exc.WorkflowInspectionError,
-            wf_svc.request,
+            workflow_service.request,
             self.get_wf_def(TEST_PACK_PATH, wf_meta),
             ac_ex_db,
             self.mock_st2_context(ac_ex_db)
@@ -182,12 +182,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Exception is expected on request of workflow execution.
         self.assertRaises(
             orquesta_exc.WorkflowInspectionError,
-            wf_svc.request,
+            workflow_service.request,
             self.get_wf_def(TEST_PACK_PATH, wf_meta),
             ac_ex_db,
             self.mock_st2_context(ac_ex_db)
@@ -199,12 +199,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Exception is expected on request of workflow execution.
         self.assertRaises(
             orquesta_exc.WorkflowInspectionError,
-            wf_svc.request,
+            workflow_service.request,
             self.get_wf_def(TEST_PACK_PATH, wf_meta),
             ac_ex_db,
             self.mock_st2_context(ac_ex_db)
@@ -215,12 +215,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Request the workflow execution.
         wf_def = self.get_wf_def(TEST_PACK_PATH, wf_meta)
         st2_ctx = self.mock_st2_context(ac_ex_db)
-        wf_ex_db = wf_svc.request(wf_def, ac_ex_db, st2_ctx)
+        wf_ex_db = workflow_service.request(wf_def, ac_ex_db, st2_ctx)
         spec_module = specs_loader.get_spec_module(wf_ex_db.spec['catalog'])
         wf_spec = spec_module.WorkflowSpec.deserialize(wf_ex_db.spec)
 
@@ -230,7 +230,7 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
         task_ctx = {'foo': 'bar'}
         st2_ctx = {'execution_id': wf_ex_db.action_execution}
 
-        task_req = {
+        task_ex_req = {
             'id': task_id,
             'spec': task_spec,
             'ctx': task_ctx,
@@ -239,7 +239,7 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
             ]
         }
 
-        wf_svc.request_task_execution(wf_ex_db, st2_ctx, task_req)
+        workflow_service.request_task_execution(wf_ex_db, st2_ctx, task_ex_req)
 
         # Check task execution is saved to the database.
         task_ex_dbs = wf_db_access.TaskExecution.query(workflow_execution=str(wf_ex_db.id))
@@ -250,7 +250,7 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
         self.assertIsNotNone(task_ex_db.id)
         self.assertGreater(task_ex_db.rev, 0)
         self.assertEqual(task_ex_db.workflow_execution, str(wf_ex_db.id))
-        self.assertEqual(task_ex_db.status, wf_lib_states.RUNNING)
+        self.assertEqual(task_ex_db.status, wf_states.RUNNING)
 
         # Check action execution for the task query with task execution ID.
         ac_ex_dbs = ex_db_access.ActionExecution.query(task_execution=str(task_ex_db.id))
@@ -265,12 +265,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Request the workflow execution.
         wf_def = self.get_wf_def(TEST_PACK_PATH, wf_meta)
         st2_ctx = self.mock_st2_context(ac_ex_db)
-        wf_ex_db = wf_svc.request(wf_def, ac_ex_db, st2_ctx)
+        wf_ex_db = workflow_service.request(wf_def, ac_ex_db, st2_ctx)
         spec_module = specs_loader.get_spec_module(wf_ex_db.spec['catalog'])
         wf_spec = spec_module.WorkflowSpec.deserialize(wf_ex_db.spec)
 
@@ -280,7 +280,7 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
         task_ctx = {'foo': 'bar'}
         st2_ctx = {'execution_id': wf_ex_db.action_execution}
 
-        task_req = {
+        task_ex_req = {
             'id': task_id,
             'spec': task_spec,
             'ctx': task_ctx,
@@ -290,11 +290,11 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
         }
 
         self.assertRaises(
-            ac_exc.InvalidActionReferencedException,
-            wf_svc.request_task_execution,
+            action_exc.InvalidActionReferencedException,
+            workflow_service.request_task_execution,
             wf_ex_db,
             st2_ctx,
-            task_req
+            task_ex_req
         )
 
     def test_handle_action_execution_completion(self):
@@ -302,12 +302,12 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Manually create the liveaction and action execution objects without publishing.
         lv_ac_db = lv_db_models.LiveActionDB(action=wf_meta['name'])
-        lv_ac_db, ac_ex_db = ac_svc.create_request(lv_ac_db)
+        lv_ac_db, ac_ex_db = action_service.create_request(lv_ac_db)
 
         # Request and pre-process the workflow execution.
         wf_def = self.get_wf_def(TEST_PACK_PATH, wf_meta)
         st2_ctx = self.mock_st2_context(ac_ex_db)
-        wf_ex_db = wf_svc.request(wf_def, ac_ex_db, st2_ctx)
+        wf_ex_db = workflow_service.request(wf_def, ac_ex_db, st2_ctx)
         wf_ex_db = self.prep_wf_ex(wf_ex_db)
 
         # Manually request task execution.
@@ -315,3 +315,34 @@ class WorkflowExecutionServiceTest(st2tests.WorkflowTestCase):
 
         # Check that a new task is executed.
         self.assert_task_running('task2')
+
+    def test_evaluate_action_execution_delay(self):
+        # No task delay.
+        task_ex_req = {'task_id': 'task1', 'task_name': 'task1'}
+        ac_ex_req = {'action': 'core.noop', 'input': None}
+        actual_delay = workflow_service.eval_action_execution_delay(task_ex_req, ac_ex_req)
+        self.assertIsNone(actual_delay)
+
+        # Simple task delay.
+        task_ex_req = {'task_id': 'task1', 'task_name': 'task1', 'delay': 180}
+        ac_ex_req = {'action': 'core.noop', 'input': None}
+        actual_delay = workflow_service.eval_action_execution_delay(task_ex_req, ac_ex_req)
+        self.assertEqual(actual_delay, 180)
+
+        # Task delay for with items task and with no concurrency.
+        task_ex_req = {'task_id': 'task1', 'task_name': 'task1', 'delay': 180, 'concurrency': None}
+        ac_ex_req = {'action': 'core.noop', 'input': None, 'items_id': 0}
+        actual_delay = workflow_service.eval_action_execution_delay(task_ex_req, ac_ex_req, True)
+        self.assertEqual(actual_delay, 180)
+
+        # Task delay for with items task, with concurrency, and evaluate first item.
+        task_ex_req = {'task_id': 'task1', 'task_name': 'task1', 'delay': 180, 'concurrency': 1}
+        ac_ex_req = {'action': 'core.noop', 'input': None, 'item_id': 0}
+        actual_delay = workflow_service.eval_action_execution_delay(task_ex_req, ac_ex_req, True)
+        self.assertEqual(actual_delay, 180)
+
+        # Task delay for with items task, with concurrency, and evaluate later items.
+        task_ex_req = {'task_id': 'task1', 'task_name': 'task1', 'delay': 180, 'concurrency': 1}
+        ac_ex_req = {'action': 'core.noop', 'input': None, 'item_id': 1}
+        actual_delay = workflow_service.eval_action_execution_delay(task_ex_req, ac_ex_req, True)
+        self.assertIsNone(actual_delay)